custom-cloud-time-processor/src/main/java/customcloudtimeprocessor/processor/DateFormatProcessor.java
package customcloudtimeprocessor.processor;

import lombok.extern.slf4j.Slf4j;

import org.springframework.cloud.stream.messaging.Processor;
import org.springframework.integration.annotation.Transformer;
import org.springframework.stereotype.Component;

import java.time.Instant;
import java.time.LocalDateTime;
import java.util.TimeZone;

@Slf4j
@Component
public class DateFormatProcessor {

    @Transformer(inputChannel = Processor.INPUT, outputChannel = Processor.OUTPUT)
    public Object transform(Long timestamp) {
        log.info("Got the input as {}", timestamp);
        return LocalDateTime.ofInstant(Instant.ofEpochMilli(timestamp),
                TimeZone.getDefault().toZoneId());
    }
}
